Addrutil: Move to snet
