+++ go/lib/snet/addrutil/BUILD.bazel
@@ -3,8 +3,8 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["addrutil.go"],
-    importpath = "github.com/scionproto/scion/go/path_srv/internal/addrutil",
-    visibility = ["//go/path_srv:__subpackages__"],
+    importpath = "github.com/scionproto/scion/go/lib/snet/addrutil",
+    visibility = ["//visibility:public"],
     deps = [
         "//go/lib/addr:go_default_library",
         "//go/lib/common:go_default_library",
