+++ go/path_srv/internal/handlers/BUILD.bazel
@@ -34,8 +34,8 @@ go_library(
         "//go/lib/scrypto:go_default_library",
         "//go/lib/scrypto/trc:go_default_library",
         "//go/lib/snet:go_default_library",
+        "//go/lib/snet/addrutil:go_default_library",
         "//go/lib/topology:go_default_library",
-        "//go/path_srv/internal/addrutil:go_default_library",
         "//go/path_srv/internal/config:go_default_library",
         "//go/path_srv/internal/segutil:go_default_library",
         "//go/proto:go_default_library",
