+++ go/path_srv/internal/segsyncer/BUILD.bazel
@@ -19,7 +19,7 @@ go_library(
         "//go/lib/periodic:go_default_library",
         "//go/lib/revcache:go_default_library",
         "//go/lib/scrypto:go_default_library",
-        "//go/path_srv/internal/addrutil:go_default_library",
+        "//go/lib/snet/addrutil:go_default_library",
         "//go/path_srv/internal/handlers:go_default_library",
         "//go/path_srv/internal/segutil:go_default_library",
         "//go/proto:go_default_library",
